requirements.txt
flask
flask-cors
boto3
requests
python-dotenv
gunicorn
